added typescript node_modules, tsconfig, tslint, and created directory bin for js files.
