+++ app/.gitignore
@@ -15,3 +15,5 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+/.jest
+/bin
