+++ app/src/modules/__tests__/thing.test.ts
@@ -1,0 +1,3 @@
+it('should work on a ts file', () => {
+  expect(2 + 2).toEqual(4);
+})
