+++ app/src/modules/thing.ts
@@ -1,0 +1,1 @@
+export const thing = "thinasdg";
